refactor(converters): one chunked CSC column sort for both routes
PR #203 and PR #204 each carried the same in-place sort of a CSC
matrix's columns: a stable argsort on column * n_rows + row over a
bounded chunk of columns at a time. csc_assembly.sort_csc_columns is
now the single implementation with the single SORT_CHUNK_ENTRIES
budget; CscAssembly._sort_columns calls it and the streaming route's
summed table imports it. The method passes the budget at call time so
the test that shrinks the module constant still sorts in chunks.

The chunk-boundary test moves next to the function, in
test_csc_assembly.py.
