+++ thyra/converters/spatialdata/csc_assembly.py
@@ -31,7 +31,8 @@ without a sort: the ``COO -> CSC`` conversion the tables used to pay for,
 one scipy call over the whole image and the single largest phase of a
 measured grid conversion, is gone rather than tuned. A source read in
 some other order gets its columns sorted afterwards, a bounded chunk of
-columns at a time, so the stored matrix is canonical either way.
+columns at a time (:func:`sort_csc_columns`, which the streaming route's
+summed table shares), so the stored matrix is canonical either way.
 
 :meth:`matrix` wraps the memmaps as a :class:`scipy.sparse.csc_matrix`
 without copying them (index dtypes are chosen up front by scipy's own
@@ -70,6 +71,69 @@ MAX_COUNT_BYTES = 1 << 30
 SORT_CHUNK_ENTRIES = 16_000_000
 
 _COUNT_DTYPE = np.uint32
+
+
+def sort_csc_columns(
+    indices: Any,
+    data: Any,
+    indptr: NDArray[Any],
+    n_rows: int,
+    chunk_entries: int = SORT_CHUNK_ENTRIES,
+) -> None:
+    """Put each column's entries in ascending row order, in place, chunk by chunk.
+
+    A scatter writes a column's entries in the order the source handed its
+    rows over. A raster read makes that ascending and the matrix canonical
+    for free. Any other order -- a TDF acquisition of several areas
+    measured one after another, whose frames come back area by area --
+    leaves the row indices within a column unsorted, which scipy reports
+    as non-canonical and which breaks every consumer that binary-searches
+    a column. The columns are already grouped, so this is a sort *within*
+    columns over a bounded slice of the arrays at a time. The assembly
+    here and the streaming route's summed table
+    (``StreamingSpatialDataConverter._convert_to_csc_no_cache``) both sort
+    their memmaps with it.
+
+    Args:
+        indices: CSC row indices, one array-like (a memmap) of the
+            non-zero count.
+        data: CSC values, aligned with ``indices``.
+        indptr: Column pointers, ``n_cols + 1`` entries of any integer
+            dtype.
+        n_rows: Row count of the matrix; every row index is below it.
+        chunk_entries: Entries a chunk may hold. A column larger than
+            this is sorted on its own.
+    """
+    indptr = np.asarray(indptr, dtype=np.int64)
+    n_cols = int(indptr.size - 1)
+    logger.info(
+        "Rows arrived out of raster order; sorting %s columns in chunks",
+        f"{n_cols:,}",
+    )
+    start_col = 0
+    while start_col < n_cols:
+        # The largest column range whose entries fit the chunk budget,
+        # and at least one column even when that column alone exceeds it.
+        end_col = int(
+            np.searchsorted(indptr, indptr[start_col] + chunk_entries, side="right")
+        )
+        end_col = max(min(end_col - 1, n_cols), start_col + 1)
+        lo, hi = int(indptr[start_col]), int(indptr[end_col])
+        if hi > lo:
+            rows = np.asarray(indices[lo:hi]).astype(np.int64)
+            column = np.repeat(
+                np.arange(start_col, end_col, dtype=np.int64),
+                np.diff(indptr[start_col : end_col + 1]),
+            )
+            # One combined key, unique because a row occurs once per
+            # column, sorted stably: the columns are already in order, so
+            # the key is a sequence of nearly sorted runs that timsort
+            # merges in a fraction of a general sort's time (measured
+            # 0.5 s against 3.1 s for lexsort on 16M entries).
+            order = np.argsort(column * n_rows + rows, kind="stable")
+            indices[lo:hi] = rows[order]
+            data[lo:hi] = np.asarray(data[lo:hi])[order]
+        start_col = end_col
 
 
 def index_dtype(n_nonzeros: int, n_rows: int) -> Any:
@@ -275,44 +339,19 @@ class CscAssembly:
     def _sort_columns(self) -> None:
         """Put each column's entries in ascending row order, chunk by chunk.
 
-        Only needed when the source handed its rows out of order; the
-        columns are already grouped, so this is a sort *within* columns
-        over a bounded slice of the arrays at a time.
+        Only needed when the source handed its rows out of order; see
+        :func:`sort_csc_columns`. The chunk budget is read at call time
+        so a test can shrink it through the module constant.
         """
         assert self.indptr is not None and self._indices is not None
         assert self._data is not None
-        indptr = self.indptr.astype(np.int64, copy=False)
-        n_cols = int(indptr.size - 1)
-        logger.info(
-            "Rows arrived out of raster order; sorting %s columns in chunks",
-            f"{n_cols:,}",
+        sort_csc_columns(
+            self._indices,
+            self._data,
+            self.indptr,
+            self.n_rows,
+            chunk_entries=SORT_CHUNK_ENTRIES,
         )
-        start_col = 0
-        while start_col < n_cols:
-            # The largest column range whose entries fit the chunk budget,
-            # and at least one column even when that column alone exceeds it.
-            end_col = int(
-                np.searchsorted(
-                    indptr, indptr[start_col] + SORT_CHUNK_ENTRIES, side="right"
-                )
-            )
-            end_col = max(min(end_col - 1, n_cols), start_col + 1)
-            lo, hi = int(indptr[start_col]), int(indptr[end_col])
-            if hi > lo:
-                rows = np.asarray(self._indices[lo:hi]).astype(np.int64)
-                column = np.repeat(
-                    np.arange(start_col, end_col, dtype=np.int64),
-                    np.diff(indptr[start_col : end_col + 1]),
-                )
-                # One combined key, unique because a row occurs once per
-                # column, sorted stably: the columns are already in order,
-                # so the key is a sequence of nearly sorted runs that
-                # timsort merges in a fraction of a general sort's time
-                # (measured 0.5 s against 3.1 s for lexsort on 16M entries).
-                order = np.argsort(column * self.n_rows + rows, kind="stable")
-                self._indices[lo:hi] = rows[order]
-                self._data[lo:hi] = np.asarray(self._data[lo:hi])[order]
-            start_col = end_col
         self.rows_in_order = True
 
     def release(self) -> None:
